refactor(core): moves katavault to the root of the src directory
