+++ packages/core/src/facades/index.ts
@@ -1,2 +1,1 @@
 export { default as AppManager } from './AppManager';
-export { default as Katavault } from './Katavault';
