+++ packages/documentation/docs/03-react/02-getting-started.mdx
@@ -77,7 +77,7 @@ See [`CreateKatavaultParameters`](04-glossary/02-types.mdx#createkatavaultparame
 
 ### 2. Authentication
 
-As explained in the [core](02-core/03-api/02-getting-started.mdx#2-authentication) documentation, before the app is able to start using the provider, the user **MUST** authenticate.
+As explained in the [core](02-core/02-getting-started.mdx#2-authentication) documentation, before the app is able to start using the provider, the user **MUST** authenticate.
 
 #### 2.1. User authentication
